+++ gens/blueprints/gens/templates/gens.html
@@ -149,7 +149,7 @@
     inputField.placeholder = inputField.value;
 
     window.regionConfig = {
-        chr: chromJs,
+        chrom: chromJs,
         start: startJs,
         end: endJs,
         genome_build: genomeBuildJs
